Safari-609.1.11.1/LayoutTests/js/dfg-put-by-val-direct-to-generic-array-expected.txt

Test that a generic array (object) accepts DFG PutByValueDirect operation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo1() is 'world' on all iterations including after DFG tier-up.
PASS foo2() is 'world2' on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
